% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_info.R
\name{cell_info}
\alias{cell_info}
\title{Construct perturbation-cell matrix}
\usage{
cell_info(count, sgrna)
}
\arguments{
\item{count}{A gene-cell matrix.}

\item{sgrna}{A perturbation data frame which contains perturbation and cell parameters.}
}
\value{
A perturbation-cell matrix
}
\description{
Construct perturbation-cell matrix. The cell is assigned with TRUE which is infected by corresponding perturbation
}
\examples{
count <- PerturbScore::count_data
sgrna <- PerturbScore::perturbation_data
ind <- cell_info(count = count, sgrna = count)
}
