% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_data.R
\docType{data}
\name{count_data}
\alias{count_data}
\title{Sample gene count matrix}
\description{
A sample gene expression count matrix from published article (Datlinger, P., Rendeiro, A., Schmidl, C. et al. Pooled CRISPR screening with single-cell transcriptome readout. Nat Methods 14, 297–301 (2017)).
}
\examples{
count <- PerturbScore::count_data
}
