% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_filtered.R
\name{count_filtered}
\alias{count_filtered}
\title{Filter genes}
\usage{
count_filtered(count, frac = 0.01, nfeatures = 2000, mt.percent = 5)
}
\arguments{
\item{count}{A gene count matrix. Before input the count data,it is recommended to finish the imputation of count matrix.}

\item{frac}{A parameter for filtering low expression genes. By default, only genes that have expression in at least that fractions of cells are kept. Default: 0.01.}

\item{nfeatures}{A parameter for extracting high-variable genes for gene co-expression network construction. By default, the logarithmic curve between mean and variance is fitted based on the loess. Then order the standardized variances and extract nfeatures genes. Default: 2000. NULL: extract all genes.}

\item{mt.percent}{A parameter for filtering cells with high expression of mitochondrial genes. Default: 5. Filtering out cells that express more than 5\% of mitochondrial genes. NULL: No cells are filtered out.}
}
\value{
A gene count matrix
}
\description{
Filter low expression genes and remove cells with high expression of mitochondiral genes.
}
\examples{
count count <- PerturbScore::count_data
count_filter <- count_filtered(count, frac = 0.01, nfeatures = 2000, mt.percent = 5)

}
