% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_corrlation.R
\name{gene_corrlation}
\alias{gene_corrlation}
\title{Compute genes correlation}
\usage{
gene_corrlation(count)
}
\arguments{
\item{count}{A gene count matrix.}
}
\value{
A gene correlation matrix
}
\description{
Calculate correlation coefficient between genes.
}
\examples{
gene_corr <- gene_corrlation(count = count_filter)
}
