% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_regression.R
\name{gene_regression}
\alias{gene_regression}
\title{Construct data for gene ridge regression and permutation}
\usage{
gene_regression(
  count,
  cell_ident,
  selected = NULL,
  ctrl,
  gene_frac = 0.01,
  mt.percent = 5,
  slot = "scale"
)
}
\arguments{
\item{count}{A gene count matrix.}

\item{cell_ident}{A perturbation-cell matrix calculated by cell_info function.}

\item{selected}{High-variable gene number. Default: 2000.}

\item{ctrl}{A parameter for control cell. Non-perturbed cells are recommended to assign NT.}

\item{gene_frac}{A parameter for filtering low expression genes. Default: 0.01.}

\item{mt.percent}{A parameter for filtering cells with high expression of mitochondrial genes. Default: 5.}

\item{slot}{A parameter for normalizing data. Default: 'scale'.}
}
\value{
A list for gene regression. Xmat represents cell identity matrix that contains cell and perturbation parameters. Ymat represents gene count matrix.
}
\description{
Construct data for gene ridge regression and permutation
}
\examples{
mat_reg <- gene_regression(count, ind, selected = 2000,ctrl="NT",
                            gene_frac = 0.01, slot = 'scale')
}
