% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paras_extraction.R
\name{paras_extraction}
\alias{paras_extraction}
\title{Extract gene correlation coefficient}
\usage{
paras_extraction(corr, do.normalization = T, abs = T)
}
\arguments{
\item{corr}{A correlation matrix calculated by gene_corrlation function.}

\item{do.normalization}{A parameter whether to  normalize betweenness centrality. Default: T.}

\item{abs}{A parameter Whether to take the absolute value of the result. Default: T}
}
\value{
A gene network parameters data frame.
}
\description{
Extract gene network parameters included normalized betweenness centrality and gene connectivity.
}
\examples{
paras <- paras_extraction(corr = gene_corr, do.normalization = T, abs = T)
}
