% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pert_feature_plot.R
\name{pert_feature_plot}
\alias{pert_feature_plot}
\title{Dot plot}
\usage{
pert_feature_plot(
  pert_score,
  num = NULL,
  features = NULL,
  perturbations = NULL
)
}
\arguments{
\item{pert_score}{Result from perturbscore function.}

\item{num}{A parameter for extracting top phenotype genes by pvalue. Default: NULL.}

\item{features}{Interested phenotype gene list. Default: NULL.}

\item{perturbations}{Interested perturbation list. Default: NULL.}
}
\value{
A dot plot
}
\description{
Dot plot for perturbation effects
}
\examples{
p <- pert_feature_plot(pert_score,num=2,features=NULL,perturbations=NULL)
p
}
