% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_data.R
\docType{data}
\name{perturbation_data}
\alias{perturbation_data}
\title{Sample dataset containing perturbations and cells}
\description{
A sample dataset from published article (Datlinger, P., Rendeiro, A., Schmidl, C. et al. Pooled CRISPR screening with single-cell transcriptome readout. Nat Methods 14, 297–301 (2017)). This dataset contains 5905 cells and 32 perturbations.
}
\examples{
sgrna <- PerturbScore::perturbation_data
}
