% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbscore.R
\name{perturbscore}
\alias{perturbscore}
\title{Calculate perturbscore}
\usage{
perturbscore(regscore_list, net_para)
}
\arguments{
\item{regscore_list}{A regression result list calculated by regression_score function.}

\item{net_para}{Gene network parameters calculated by fc_extraction}
}
\value{
A result for perturbscore
}
\description{
Use ridge regression and gene network parameters to calculate the actual perturbation effect in single-cell CRISPR screening data.
}
\examples{
pert_score <- perturbscore(regscore_list = reg_score, net_para = paras)
}
