% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_score.R
\name{regression_score}
\alias{regression_score}
\title{Calculate perturbation effect}
\usage{
regression_score(
  count,
  sgrna,
  ctrl,
  nfeatures = 2000,
  n_permut = NULL,
  paramt = 0.01,
  filt = 0.01,
  mt.percent = 5,
  slot = "scale"
)
}
\arguments{
\item{count}{A gene count matrix.}

\item{sgrna}{A perturbation-cell data frame. Noted that the data frame contain perturbation and cell parameters.}

\item{ctrl}{A parameter for non-perturbed cells which is recommended as NT.}

\item{nfeatures}{High-variable gene number. Default: 2000.}

\item{n_permut}{The number of permutations for p value. Default: NULL.}

\item{paramt}{The Regularization coefficient. Default: 0.01.}

\item{filt}{A parameter for filtering low expression genes. Default: 0.01.}

\item{mt.percent}{A parameter for filtering cells with high expression of mitochondrial genes. Default: 5.}

\item{slot}{A parameter for normalizing data. Default: 'scale'.}
}
\value{
A list for regression result.
}
\description{
Calculate perturbation effect by ridge regression in single-cell CRISPR screening data.
}
\examples{
count <- PerturbScore::count_data
sgrna <- PerturbScore::perturbation_data
reg_score <- regression_score(count,sgrna,ctr = "NT", nfeatures = 2000,n_permut = 1000,paramt = 0.01,filt = 0.01,mt.percent = 5, slot = 'scale')
}
